# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Record Shop",
        "description": "An example application for a Record Shop selling Oldies But Goldies records.",
        "contact": {
            "email": "ismail.atkurt@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/artists": {
            "get": {
                "tags": [
                    "artists"
                ],
                "summary": "Get all artists",
                "description": "Returns all artists available",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "searchTerm",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of artists"
                    }
                }
            },
            "post": {
                "tags": [
                    "artists"
                ],
                "summary": "Create new artist",
                "description": "Create new artist with given data",
                "operationId": "create",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the artist",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Edith Piaf"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns Created Artist"
                    }
                }
            }
        },
        "/artists/{artistId}": {
            "get": {
                "tags": [
                    "artists"
                ],
                "summary": "Get single artist",
                "description": "Returns the artist with given ID.",
                "operationId": "one",
                "parameters": [
                    {
                        "name": "artistId",
                        "in": "path",
                        "description": "The id of the artist to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns Single artist"
                    }
                }
            },
            "put": {
                "tags": [
                    "artists"
                ],
                "summary": "Update artist",
                "description": "Update the artist with given ID and provided post data.",
                "operationId": "update",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Updated name of the artist",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Daft Punk"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns Updated Artist"
                    }
                }
            }
        },
...
    }
}